+++ ChangeLog.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 ## 1.0
+- FIX : Compatibility V16 TOKEN - *13/06/2022* - 1.0.7  
 - FIX : Compatibility V15 : variables inutilisées supprimées car provoques des warning
 - FIX: v14 compatibility: `f.tva` (now `total_tva`) and `f.total` (now `total_ht`)
   \+ module descriptor - *13/10/2021* - 1.0.5